pyproject.toml
[project]
name = "movielens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.35.1",
    "chromadb>=0.5.20",
    "cohere>=5.11.4",
    "fastapi[standard]>=0.115.5",
    "google-generativeai>=0.8.3",
    "ipykernel>=6.29.5",
    "moviepy>=2.0.0",
    "numpy>=2.1.3",
    "openai>=1.55.0",
    "pip>=24.3.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentence-transformers>=3.3.1",
    "sounddevice>=0.5.1",
    "streamlit>=1.40.1",
    "torchvision>=0.20.1",
    "transformers>=4.46.3",
    "uuid>=1.30",
    "uvicorn>=0.32.1",
    "wavio>=0.0.9",
]
